--- interval-timer/app/page.tsx
@@ -1,6 +1,0 @@
-import React from 'react';
-
-export default function HomePage ()
-{
-    return (<h1>Home</h1>);
-}
