+++ intervaltimer/app/(root)/page.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function Discover() {
+  return (
+    <h1>Discover</h1>
+  )
+}
