+++ intervaltimer/next.config.js
