+++ intervaltimer/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "interval-timer",
+  "name": "intervaltimer",
   "version": "0.1.0",
   "private": true,
   "scripts": {
@@ -13,19 +13,14 @@
     "@types/react": "18.2.20",
     "@types/react-dom": "18.2.7",
     "autoprefixer": "10.4.15",
-    "class-variance-authority": "^0.7.0",
-    "clsx": "^2.0.0",
     "eslint": "8.47.0",
     "eslint-config-next": "13.4.19",
-    "lucide-react": "^0.268.0",
     "next": "13.4.19",
     "pocketbase": "^0.16.0",
     "postcss": "8.4.28",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "tailwind-merge": "^1.14.0",
     "tailwindcss": "3.3.3",
-    "tailwindcss-animate": "^1.0.6",
     "typescript": "5.1.6",
     "zustand": "^4.4.1"
   }
